Removing requestLine from HttpURLConnection.getRequestProperties()
This was breaking an Android test and appears not to be part of the spec for
HttpURLConnection.

To fix the test fully another change was required to add additional headers
passed to the CookieHandler.

Also added are some tests ported over from Android that test
CookieHandler/CookieManager and improve coverage of this area.