Move all socket connection work into Connection
Motivation: Moving route selection into Connection
will, with further changes, allow the Connection to
communicate SSLSocket information to the RouteSelector
that can be used to impact future Routes to try (e.g.
it will know not to bother trying to connect again
if the connection spec contains unsupported
protocols).

The change also has another side effect: Errors during
socket creation / handshake, and errors that occur
afterward (during request write / response read) are
now easy to distinguish. In future, this would enable
OkHttp to distinguish between a handshake error, and a
problem streaming the response headers, which I believe
should be handled differently.

Before and after this change, in the event of a problem
reading the response OkHttp might (re)submit the request
body, over a different Route (e.g. with a lower protocol
version), which seems undesirable.

With this commit the existing behavior has been kept:
questions remain about how many times a body resubmission
should be retried (if any). The continued use of the
RouteSelector currently puts an upper bound on the number
of retries that might be attempted. If non-handshake
retries are retained but RouteSelector is not used then
some mechanism to provide an upper bound for retried would
have be coded.

This change has an API impact:

1) RouteSelector has been made an interface and moved into
the okhttp route package because it is now required by
Connection; Connection now takes a RouteSelector during
construction.

2) The Connection constructor now takes a boolean to indicate
whether it should retry on connection failure. This could
be removed if the logic for whether to retry was moved into
the RouteSelector (either by modifying RouteSelectorImpl
or creating "no retries" RouteSelector).

3) Connection now throws a custom exception from connect() to
indicate the connection attempt was unsuccessful. Crucially,
this does not extend IOException so it is easy to distinguish
from IOExceptions that might occur during the request write
/response read. The new exception includes information that
can be used during debugging.