(Hopefully) fix a flake in Spdy3ConnectionTest on Android
In some cases serverSocket is found to be null in
readAndWriteFrames().

The suspicion is that it's because of a lack of synchronization
between the thread setting the field, and the thread (started in
play()) reading it. serverSocket / port have been made
volatile to ensure the field state is flushed / read and not
cached. No flakes seen in several runs since the change.