Make matching of TlsVersion optionally stricter
On Android there are 4 connection specs configured:
1) TLSv1.2, TLSv1.1, TLSv1.0, SSLv3
2) TLSv1.1, TLSv1.0, SSLv3
3) TLSv1.0, SSLv3
4) SSLv3

(Typically spec 4 is not tried because SSLv3 is not
enabled by default).

If only TLSv1.0 is enabled then OkHttp will
attempt 3 connections (1, 2 and 3 above), with exactly
the same enabled protocols enabled (just TLSv1.0).

This change adds an addition restriction: the first
protocol listed *must* be supported for the connection
negotiation to be attempted so only spec 3 would be
attempted after this change.

Note, this does not change default behavior.

OkHttp by default tries twice (MODERN_TLS + COMPATIBLE_TLS).
If only TLSv1.0 is enabled: it will still try to connect
twice with TLSv1.0.