Remove IOException throws from source() and close() on ResponseBody.
source() throwing was the result of legacy integrations which now eagerly open their streams. This eliminates the need for this method to throw as well as a few downstream methods like byteStream() and charStream().

close() throwing puts the caller in an awkward position. They have already read the body (or are ignoring it) and just want to ensure resource cleanup happens. If this fails, however, and the IOException was throw there is no reasonable action for them to take other than ignoring. As such, we take the liberty of ignoring it for them.